# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  internal_api.AccountResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        example: CNY
        type: string
      id:
        example: 1
        type: integer
      initial_balance:
        example: 1000
        type: number
      name:
        example: Cash
        type: string
      type:
        example: cash
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  internal_api.AttachmentResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      file_name:
        example: attachments/1/1_photo.jpg
        type: string
      file_size:
        example: 102400
        type: integer
      height:
        example: 1080
        type: integer
      id:
        example: 1
        type: integer
      original_name:
        example: photo.jpg
        type: string
      sort_order:
        example: 0
        type: integer
      transaction_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      width:
        example: 1920
        type: integer
    type: object
...
  internal_api.CategoryResponse:
    properties:
      community_icon_id:
        example: community_123
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      custom_icon_path:
        example: /custom/icons/food.svg
        type: string
      icon:
        example: restaurant
        type: string
      icon_type:
        example: material
        type: string
      id:
        example: 1
        type: integer
      kind:
        example: expense
        type: string
      level:
        example: 2
        type: integer
      name:
        example: Food
        type: string
      parent_id:
        example: 1
        type: integer
      sort_order:
        example: 0
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  internal_api.CreateTransactionRequest:
    properties:
      account_id:
        example: 1
        type: integer
      amount:
        example: 100.5
        type: number
      category_id:
        example: 1
        type: integer
      happened_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      ledger_id:
        example: 1
        type: integer
      note:
        example: Lunch
        type: string
      to_account_id:
        example: 2
        type: integer
      type:
        enum:
        - expense
        - income
        - transfer
        example: expense
        type: string
    required:
    - amount
    - happened_at
    - ledger_id
    - type
    type: object
  internal_api.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: bad request
        type: string
      message:
        example: Invalid input
        type: string
    type: object
...
  internal_api.TagResponse:
    properties:
      color:
        example: '#FF5722'
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Business
        type: string
      sort_order:
        example: 0
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  internal_api.TransactionResponse:
    properties:
      account:
        $ref: '#/definitions/internal_api.AccountResponse'
      account_id:
        example: 1
        type: integer
      amount:
        example: 100.5
        type: number
      attachments:
        items:
          $ref: '#/definitions/internal_api.AttachmentResponse'
        type: array
      category:
        $ref: '#/definitions/internal_api.CategoryResponse'
      category_id:
        example: 1
        type: integer
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      happened_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      ledger_id:
        example: 1
        type: integer
      note:
        example: Lunch
        type: string
      recurring_id:
        type: integer
      tags:
        items:
          $ref: '#/definitions/internal_api.TagResponse'
        type: array
      to_account:
        $ref: '#/definitions/internal_api.AccountResponse'
      to_account_id:
        example: 2
        type: integer
      type:
        example: expense
        type: string
      updated_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  internal_api.UpdateTransactionRequest:
    properties:
      account_id:
        example: 1
        type: integer
      amount:
        example: 200
        type: number
      category_id:
        example: 2
        type: integer
      happened_at:
        example: "2024-01-02T12:00:00Z"
        type: string
      note:
        example: Updated note
        type: string
      to_account_id:
        example: 2
        type: integer
      type:
        enum:
        - expense
        - income
        - transfer
        example: income
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: BeeCount Cloud API 是一个用于管理个人财务账本的云服务 API，提供账本、交易、分类、标签等功能。
  license:
    name: AGPL 3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: http://swagger.io/terms/
  title: BeeCount Cloud API
  version: "1.0"
paths:
...
  /api/v1/ledgers/{ledger_id}/transactions:
    get:
      consumes:
      - application/json
      description: 返回指定账本中的交易列表，支持分页
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/internal_api.TransactionResponse'
            type: array
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取交易列表
      tags:
      - 交易
    post:
      consumes:
      - application/json
      description: 为指定账本创建新的交易记录
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      - description: 交易信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_api.TransactionResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 创建交易
      tags:
      - 交易
  /api/v1/ledgers/{ledger_id}/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定交易，只有交易所有者可以删除
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 删除成功
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 交易不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除交易
      tags:
      - 交易
    get:
      consumes:
      - application/json
      description: 根据交易ID返回交易详细信息
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/internal_api.TransactionResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 交易不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取交易详情
      tags:
      - 交易
    put:
      consumes:
      - application/json
      description: 更新指定交易的信息，只有交易所有者可以更新
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_api.TransactionResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 交易不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新交易
      tags:
      - 交易
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"